Move Budget EDSL into its own module
